refactor: move computed handling
